requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.0
pandas>=1.5.0
Pillow>=9.0.0
matplotlib>=3.6.0
seaborn>=0.12.0
scikit-learn>=1.1.0
albumentations>=1.3.0
pytorch-grad-cam>=1.4.0
